requirements.txt
sqlparse
openai
python-dotenv